+++ src/App.jsx
@@ -1,22 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-
 import './index.css'
+import TodoList from './TodoList'
+import AddToDoForm from './AddTodoForm'
 
-const todoList = [
-  {
-    objectID: 1,
-    title: "Create a repo on GitHub"
-  },
-  {
-    objectID: 2,
-    title: "Create the todo application"
-  },
-  {
-    objectID: 3,
-    title: "Complete assigment"
-  }
-]
+
 
 function App() {
   
@@ -25,17 +13,19 @@ function App() {
     
       <div>
         <h1>Todo list</h1>
-       
+
+ {/* Using AddToDoForm component*/}  
+ <AddToDoForm />
+
+{/* Using TodoList component*/}
+  <TodoList /> 
+
+      
+    
       
       <hr/>
 
-      <ul>
-        {todoList.map(function(item) {
-          return (
-            <li key={item.objectID}>{item.title}</li>
-          )
-        })}
-      </ul>
+     
       </div> 
   )
 }
